- Added RSA encrypt/decrypt - Fixed incorrect usage of size_t
git-svn-id: file:///H/Logs/temp/OpenSSL.Core_git/branches/pipelines@77 f149d0f5-08f7-8743-b74d-e9091c64294d

+++ src/OpenSSL.Core/Interop/MemoryTracker.cs
@@ -186,7 +186,7 @@ namespace OpenSSL.Core.Interop
 			return problems;
 		}
 
-		static IntPtr malloc(uint num, IntPtr file, int line)
+		static IntPtr malloc(ulong num, IntPtr file, int line)
 		{
 			lock (_memory)
 			{
@@ -195,7 +195,7 @@ namespace OpenSSL.Core.Interop
                     file = Native.PtrToStringAnsi(file, false),
                     line = line,
                     stack = new StackTrace(true),
-                    bytes = num,
+                    bytes = (uint)num,
                     ptr = Marshal.AllocHGlobal((int)num),
                     count = _tracking ? 0 : 1
 				};
@@ -226,7 +226,7 @@ namespace OpenSSL.Core.Interop
 			}
 		}
 
-		static IntPtr realloc(IntPtr addr, uint num, IntPtr file, int line)
+		static IntPtr realloc(IntPtr addr, ulong num, IntPtr file, int line)
 		{
 			lock (_memory)
 			{
@@ -238,7 +238,7 @@ namespace OpenSSL.Core.Interop
 					stack = new StackTrace(true),
 					file = Native.PtrToStringAnsi(file, false),
 					line = line,
-					bytes = num,
+					bytes = (uint)num,
 					ptr = Marshal.ReAllocHGlobal(addr, (IntPtr)num),
                     count = _tracking ? 0 : 1
                 };
